LayoutTests/platform/glib/imported/w3c/web-platform-tests/cookies/samesite/sandbox-iframe-subresource.https-expected.txt



FAIL SameSite cookies with intervening sandboxed iframe and subresources assert_equals: `samesite_lax=0.07583713055641716` in request to `https://web-platform.test:9443`. expected (undefined) undefined but got (string) "0.07583713055641716"
